move card_assets file to project root
